inmemory saga log moved to the orchestrator project
