+++ Orchestrator/Prognetics.Saga.Orchestrator.Contract/ISagaLog.cs
@@ -1,22 +1,8 @@
 using Prognetics.Saga.Orchestrator.Contract.DTO;
-using System.Collections.Concurrent;
 
 namespace Prognetics.Saga.Orchestrator.Contract;
 public interface ISagaLog
 {
     Task<TransactionState?> GetState(string transactionId);
     Task SetState(TransactionState transactionState);
-}
-
-public class SagaLog : ISagaLog
-{
-    private readonly ConcurrentDictionary<string, TransactionState> _transactions = new();
-    public Task<TransactionState?> GetState(string transactionId)
-        => Task.FromResult(_transactions.GetValueOrDefault(transactionId));
-
-    public Task SetState(TransactionState transactionState)
-    {
-        _transactions[transactionState.TransactionId] = transactionState;
-        return Task.CompletedTask;
-    }
 }
